AGREG-3621 add default sensors for cpp wrapper wait for async start& stop
