+++ src/wrapper/DataCollector.cpp
@@ -2,304 +2,12 @@
 
 #include "DataCollector.h"
 #include "DataCollectorImpl.h"
-#include "HSMSensorImpl.h"
-#include "HSMBarSensorImpl.h"
-#include "HSMLastValueSensorImpl.h"
-#include "HSMBaseParamsFuncSensor.h"
-#include "HSMParamsFuncSensorImpl.h"
-#include "HSMSensorOptionsImpl.h"
 
-#include "msclr/auto_gcroot.h"
 #include "msclr/marshal_cppstd.h"
-
-
-using namespace HSMDataCollector::Core;
-using namespace HSMDataCollector::PublicInterface;
-using namespace HSMDataCollector::Logging;
-using namespace HSMDataCollector::Options;
-using namespace HSMDataCollector::Alerts;
-using namespace HSMSensorDataObjects::SensorRequests;
-using HSMSensorDataObjects::SensorStatus;
-using System::String;
-using System::Func;
-using System::Collections::Generic::List;
-using System::Threading::Tasks::Task;
 
 using namespace std;
 using namespace hsm_wrapper;
 
-DataCollectorImpl::DataCollectorImpl(const std::string& product_key, const std::string& address, int port, const string& module)
-{
-	CollectorOptions^ options = gcnew CollectorOptions();
-	options->AccessKey = gcnew String(product_key.c_str());
-	options->ServerAddress = gcnew String(address.c_str());
-	options->Port = port;
-	options->Module = gcnew String(module.c_str());
-	data_collector = gcnew DataCollector(options);
-}
-
-DataCollectorImpl::~DataCollectorImpl()
-{
-	if (!!start_task)
-	{
-		start_task->GetAwaiter().GetResult();
-	}
-}
-
-void DataCollectorImpl::Initialize(const std::string& config_path, bool write_debug)
-{
-	LoggerOptions^ options = gcnew LoggerOptions();
-	options->WriteDebug = write_debug;
-	if (!config_path.empty())
-		options->ConfigPath = gcnew String(config_path.c_str());
-
-	data_collector->AddNLog(options);
-}
-
-void DataCollectorImpl::Start()
-{
-	data_collector->Start();
-}
-
-void DataCollectorImpl::StartAsync()
-{
-	start_task = Task::Run(gcnew System::Func<Task^>(data_collector.get(), &IDataCollector::Start));
-}
-
-void DataCollectorImpl::Stop()
-{
-	data_collector->Stop();
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeSystemMonitoring(bool is_cpu, bool is_free_ram)
-{
-	if (is_cpu) data_collector->Windows->AddTotalCpu();
-	if (is_free_ram) data_collector->Windows->AddFreeRamMemory();
-}
-
-void DataCollectorImpl::InitializeDiskMonitoring(const string& target, bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
-{
-	auto disk_options = gcnew HSMDataCollector::Options::DiskSensorOptions();
-	disk_options->TargetPath = gcnew System::String(target.c_str());
-
-	if (is_free_space) data_collector->Windows->AddFreeDiskSpace(disk_options);
-	if (is_free_space_prediction) data_collector->Windows->AddFreeDisksSpacePrediction(disk_options);
-
-	auto disk_bar_options = gcnew HSMDataCollector::Options::DiskBarSensorOptions();
-	disk_bar_options->TargetPath = gcnew System::String(target.c_str());
-
-	if (is_active_time) data_collector->Windows->AddActiveDisksTime(disk_bar_options);
-	if (is_queue_lenght) data_collector->Windows->AddDisksQueueLength(disk_bar_options);
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeAllDisksMonitoring(bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
-{
-	if (is_free_space) data_collector->Windows->AddFreeDisksSpace();
-	if (is_free_space_prediction) data_collector->Windows->AddFreeDisksSpacePrediction();
-	if (is_active_time) data_collector->Windows->AddActiveDisksTime();
-	if (is_queue_lenght) data_collector->Windows->AddDisksQueueLength();
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeProcessMonitoring(bool is_cpu, bool is_memory, bool is_threads)
-{
-	if (is_cpu) data_collector->Windows->AddProcessCpu();
-	if (is_memory) data_collector->Windows->AddProcessMemory();
-	if (is_threads) data_collector->Windows->AddProcessThreadCount();
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeOsMonitoring(bool is_last_update, bool is_last_restart)
-{
- 	if (is_last_update) data_collector->Windows->AddWindowsLastUpdate();
- 	if (is_last_restart) data_collector->Windows->AddWindowsLastRestart();
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeOsLogsMonitoring(bool is_warning, bool is_error)
-{
-	if (is_warning) data_collector->Windows->AddWarningWindowsLogs();
-	if (is_error) data_collector->Windows->AddErrorWindowsLogs();
-}
-
-
-void hsm_wrapper::DataCollectorImpl::InitializeProductVersion(const std::string& version)
-{
-	VersionSensorOptions^ options = gcnew VersionSensorOptions();
-	options->Version = gcnew System::Version(gcnew String(version.c_str()));
-	data_collector->Windows->AddProductVersion(options);
-}
-
-void hsm_wrapper::DataCollectorImpl::InitializeCollectorMonitoring(bool is_alive /*= true*/, bool version /*= true*/, bool status /*= true*/)
-{
-	if (is_alive) data_collector->Windows->AddCollectorAlive();
-	if (version) data_collector->Windows->AddCollectorVersion();
-}
-
-void hsm_wrapper::DataCollectorImpl::AddServiceStateMonitoring(const std::string& service_name)
-{
-	data_collector->Windows->SubscribeToWindowsServiceStatus(gcnew String(service_name.c_str()));
-}
-
-void hsm_wrapper::DataCollectorImpl::SendFileAsync(const std::string& sensor_path, const std::string& file_path, HSMSensorStatus status /*= HSMSensorStatus::Ok*/, const std::string& description /*= {}*/)
-{
-	auto task = data_collector->SendFileAsync(gcnew String(sensor_path.c_str()), gcnew String(file_path.c_str()), SensorStatus{ status }, gcnew String(description.c_str()));
-}
-
-namespace {
-
-	template <class T, class A>
-	T^ ConvertAlert(const A& alert)
-	{
-		auto hsm_alert = alert.Impl()->GetAlert<T>();
-		return hsm_alert;
-	}
-
-	template <class T, class A>
-	List<T^>^ ConvertAlerts(const vector<A>& alerts)
-	{
-		auto hsm_alerts = gcnew List<T^>();
-
-		for (auto& alert : alerts)
-		{
-			auto hsm_alert = ConvertAlert<T>(alert);
-			hsm_alerts->Add(hsm_alert);
-		}
-		return hsm_alerts;
-	}
-
-	void ConvertsBaseOptions(HSMDataCollector::Options::SensorOptions^ hsm_options, const hsm_wrapper::HSMBaseSensorOptions& options)
-	{
-		hsm_options->Description = gcnew String(options.description.c_str());
-		if (options.keep_history)	hsm_options->KeepHistory = TimeSpan::FromMilliseconds(options.keep_history->count());
-		if (options.self_destroy)	hsm_options->SelfDestroy = TimeSpan::FromMilliseconds(options.self_destroy->count());
-		if (options.ttl)			hsm_options->TTL		 = TimeSpan::FromMilliseconds(options.ttl->count());
-		if (options.enable_for_grafana)		hsm_options->EnableForGrafana = options.enable_for_grafana.value();
-		if (options.is_singleton_sensor)	hsm_options->IsSingletonSensor = options.is_singleton_sensor.value();
-		if (options.aggregate_data)			hsm_options->AggregateData = options.aggregate_data.value();
-	}
-
-	HSMDataCollector::Options::InstantSensorOptions^ ConvertInstantOptions(const hsm_wrapper::HSMInstantSensorOptions& options)
-	{
-		auto hsm_options = gcnew InstantSensorOptions();
-
-		ConvertsBaseOptions(hsm_options, options);
-		
-		hsm_options->Alerts = ConvertAlerts<InstantAlertTemplate>(options.alerts);
-
-		return hsm_options;
-	}
-
-	HSMDataCollector::Options::BarSensorOptions^ ConvertBarOptions(const hsm_wrapper::HSMBarSensorOptions& options)
-	{
-		auto hsm_options = gcnew HSMDataCollector::Options::BarSensorOptions();
-
-		ConvertsBaseOptions(hsm_options, options);
-
-		hsm_options->BarPeriod = TimeSpan::FromMilliseconds(options.bar_period.count());
-		hsm_options->BarTickPeriod = TimeSpan::FromMilliseconds(options.bar_tick_period.count());
-		hsm_options->PostDataPeriod = TimeSpan::FromMilliseconds(options.post_data_period.count());
-		hsm_options->Precision = options.precision;
-		hsm_options->Alerts = ConvertAlerts<BarAlertTemplate>(options.alerts);
-
-		return hsm_options;
-	}
-}
-
-HSMSensor<bool> DataCollectorImpl::CreateBoolSensor(const std::string& path, const std::string& description)
-{
-	auto bool_sensor = data_collector->CreateBoolSensor(gcnew String(path.c_str()), gcnew String(description.c_str()));
-	return HSMSensor<bool>{std::make_shared<HSMSensorImpl<bool>>(bool_sensor)};
-}
-
-HSMSensor<bool> DataCollectorImpl::CreateBoolSensor(const std::string& path, const HSMInstantSensorOptions& options)
-{
-	auto bool_sensor = data_collector->CreateBoolSensor(gcnew String(path.c_str()), ConvertInstantOptions(options));
-	return HSMSensor<bool>{std::make_shared<HSMSensorImpl<bool>>(bool_sensor)};
-}
-
-HSMSensor<int> DataCollectorImpl::CreateIntSensor(const std::string& path, const std::string& description)
-{
-	auto int_sensor = data_collector->CreateIntSensor(gcnew String(path.c_str()), gcnew String(description.c_str()));
-	return HSMSensor<int>{std::make_shared<HSMSensorImpl<int>>(int_sensor)};
-}
-
-HSMSensor<int> DataCollectorImpl::CreateIntSensor(const std::string& path, const hsm_wrapper::HSMInstantSensorOptions& options)
-{
-	auto int_sensor = data_collector->CreateIntSensor(gcnew String(path.c_str()), ConvertInstantOptions(options));
-	return HSMSensor<int>{std::make_shared<HSMSensorImpl<int>>(int_sensor)};
-}
-
-HSMSensor<double> DataCollectorImpl::CreateDoubleSensor(const std::string& path, const std::string& description)
-{
-	auto double_sensor = data_collector->CreateDoubleSensor(gcnew String(path.c_str()), gcnew String(description.c_str()));
-	return HSMSensor<double>{std::make_shared<HSMSensorImpl<double>>(double_sensor)};
-}
-
-HSMSensor<double> DataCollectorImpl::CreateDoubleSensor(const std::string& path, const hsm_wrapper::HSMInstantSensorOptions& options)
-{
-	auto double_sensor = data_collector->CreateDoubleSensor(gcnew String(path.c_str()), ConvertInstantOptions(options));
-	return HSMSensor<double>{std::make_shared<HSMSensorImpl<double>>(double_sensor)};
-}
-
-HSMSensor<string> DataCollectorImpl::CreateStringSensor(const std::string& path, const std::string& description)
-{
-	auto string_sensor = data_collector->CreateStringSensor(gcnew String(path.c_str()), gcnew String(description.c_str()));
-	return HSMSensor<string>{std::make_shared<HSMSensorImpl<string>>(string_sensor)};
-}
-
-HSMSensor<string> DataCollectorImpl::CreateStringSensor(const std::string& path, const hsm_wrapper::HSMInstantSensorOptions& options)
-{
-	auto string_sensor = data_collector->CreateStringSensor(gcnew String(path.c_str()), ConvertInstantOptions(options));
-	return HSMSensor<string>{std::make_shared<HSMSensorImpl<string>>(string_sensor)};
-}
-
-#ifdef ENABLE_OBSOLETE
-HSMLastValueSensor<bool> DataCollectorImpl::CreateLastValueBoolSensor(const std::string& path, bool default_value, const std::string& description)
-{
-	auto int_default_sensor = data_collector->CreateLastValueBoolSensor(gcnew String(path.c_str()), default_value, gcnew String(description.c_str()));
-	return HSMLastValueSensor<bool>{std::make_shared<HSMLastValueSensorImpl<bool>>(int_default_sensor)};
-}
-
-HSMLastValueSensor<int> DataCollectorImpl::CreateLastValueIntSensor(const std::string& path, int default_value, const std::string& description)
-{
-	auto int_default_sensor = data_collector->CreateLastValueIntSensor(gcnew String(path.c_str()), default_value, gcnew String(description.c_str()));
-	return HSMLastValueSensor<int>{std::make_shared<HSMLastValueSensorImpl<int>>(int_default_sensor)};
-}
-
-HSMLastValueSensor<double> DataCollectorImpl::CreateLastValueDoubleSensor(const std::string& path, double default_value, const std::string& description)
-{
-	auto double_default_sensor = data_collector->CreateLastValueDoubleSensor(gcnew String(path.c_str()), default_value, gcnew String(description.c_str()));
-	return HSMLastValueSensor<double>{std::make_shared<HSMLastValueSensorImpl<double>>(double_default_sensor)};
-}
-
-HSMLastValueSensor<std::string> DataCollectorImpl::CreateLastValueStringSensor(const std::string& path, const std::string& default_value, const std::string& description)
-{
-	auto int_default_sensor = data_collector->CreateLastValueStringSensor(gcnew String(path.c_str()), gcnew String(default_value.c_str()), gcnew String(description.c_str()));
-	return HSMLastValueSensor<string>{std::make_shared<HSMLastValueSensorImpl<std::string>>(int_default_sensor)};
-}
-#endif
-
-HSMBarSensor<int> DataCollectorImpl::CreateIntBarSensor(const std::string& path, int timeout, int small_period, const std::string& description)
-{
-	auto int_bar_sensor = data_collector->CreateIntBarSensor(gcnew String(path.c_str()), timeout, small_period, gcnew String(description.c_str()));
-	return HSMBarSensor<int>{std::make_shared<HSMBarSensorImpl<int>>(int_bar_sensor)};
-}
-
-HSMBarSensor<int> DataCollectorImpl::CreateIntBarSensor(const std::string& path, const HSMBarSensorOptions& options)
-{
-	auto int_bar_sensor = data_collector->CreateIntBarSensor(gcnew String(path.c_str()), ConvertBarOptions(options));
-	return HSMBarSensor<int>{std::make_shared<HSMBarSensorImpl<int>>(int_bar_sensor)};
-}
-
-HSMBarSensor<double> DataCollectorImpl::CreateDoubleBarSensor(const std::string& path, int timeout, int small_period, int precision, const std::string& description)
-{
-	auto double_bar_sensor = data_collector->CreateDoubleBarSensor(gcnew String(path.c_str()), timeout, small_period, precision, gcnew String(description.c_str()));
-	return HSMBarSensor<double>{std::make_shared<HSMBarSensorImpl<double>>(double_bar_sensor)};
-}
-
-HSMBarSensor<double> DataCollectorImpl::CreateDoubleBarSensor(const std::string& path, const HSMBarSensorOptions& options)
-{
-	auto double_bar_sensor = data_collector->CreateDoubleBarSensor(gcnew String(path.c_str()), ConvertBarOptions(options));
-	return HSMBarSensor<double>{std::make_shared<HSMBarSensorImpl<double>>(double_bar_sensor)};
-}
 
 DataCollectorImplWrapper::DataCollectorImplWrapper(const std::string& product_key, const std::string& address, int port, const std::string& module) 
 try : impl( new DataCollectorImpl(product_key, address, port, module))
@@ -358,11 +66,11 @@ void DataCollectorImplWrapper::StartAsync()
 	}
 }
 
-void hsm_wrapper::DataCollectorImplWrapper::InitializeSystemMonitoring(bool is_cpu, bool is_free_ram)
+void DataCollectorImplWrapper::StopAsync()
 {
 	try
 	{
-		impl->InitializeSystemMonitoring(is_cpu, is_free_ram);
+		impl->StopAsync();
 	}
 	catch (System::Exception^ ex)
 	{
@@ -370,11 +78,11 @@ void hsm_wrapper::DataCollectorImplWrapper::InitializeSystemMonitoring(bool is_c
 	}
 }
 
-void hsm_wrapper::DataCollectorImplWrapper::InitializeCollectorMonitoring(bool is_alive, bool version, bool status) 
+void hsm_wrapper::DataCollectorImplWrapper::InitializeSystemMonitoring(bool is_cpu, bool is_free_ram, bool is_time_in_gc)
 {
 	try
 	{
-		impl->InitializeCollectorMonitoring(is_alive, version);
+		impl->InitializeSystemMonitoring(is_cpu, is_free_ram, is_time_in_gc);
 	}
 	catch (System::Exception^ ex)
 	{
@@ -382,11 +90,11 @@ void hsm_wrapper::DataCollectorImplWrapper::InitializeCollectorMonitoring(bool i
 	}
 }
 
-void hsm_wrapper::DataCollectorImplWrapper::InitializeProcessMonitoring(bool is_cpu, bool is_memory, bool is_threads)
+void hsm_wrapper::DataCollectorImplWrapper::InitializeCollectorMonitoring(bool is_alive, bool is_version, bool is_errors) 
 {
 	try
 	{
-		impl->InitializeProcessMonitoring(is_cpu, is_memory, is_threads);
+		impl->InitializeCollectorMonitoring(is_alive, is_version, is_errors);
 	}
 	catch (System::Exception^ ex)
 	{
@@ -394,11 +102,23 @@ void hsm_wrapper::DataCollectorImplWrapper::InitializeProcessMonitoring(bool is_
 	}
 }
 
-void hsm_wrapper::DataCollectorImplWrapper::InitializeOsMonitoring(bool is_last_update, bool is_last_restart)
+void hsm_wrapper::DataCollectorImplWrapper::InitializeProcessMonitoring(bool is_cpu, bool is_memory, bool is_threads, bool is_time_in_gc)
 {
 	try
 	{
-		impl->InitializeOsMonitoring(is_last_update, is_last_restart);
+		impl->InitializeProcessMonitoring(is_cpu, is_memory, is_threads, is_time_in_gc);
+	}
+	catch (System::Exception^ ex)
+	{
+		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
+	}
+}
+
+void hsm_wrapper::DataCollectorImplWrapper::InitializeOsMonitoring(bool is_last_update, bool is_last_restart, bool is_version)
+{
+	try
+	{
+		impl->InitializeOsMonitoring(is_last_update, is_last_restart, is_version);
 	}
 	catch (System::Exception^ ex)
 	{
@@ -430,11 +150,11 @@ void hsm_wrapper::DataCollectorImplWrapper::AddServiceStateMonitoring(const std:
 	}
 }
 
-void DataCollectorImplWrapper::InitializeDiskMonitoring(const string& target, bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
+void DataCollectorImplWrapper::InitializeQueueDiagnostic(bool is_overflow, bool is_process_time, bool is_values_count, bool is_content_size)
 {
 	try
 	{
-		impl->InitializeDiskMonitoring(target, is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght);
+		impl->InitializeQueueDiagnostic(is_overflow, is_process_time, is_values_count, is_content_size);
 	}
 	catch (System::Exception^ ex)
 	{
@@ -442,11 +162,35 @@ void DataCollectorImplWrapper::InitializeDiskMonitoring(const string& target, bo
 	}
 }
 
-void DataCollectorImplWrapper::InitializeAllDisksMonitoring(bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
+void hsm_wrapper::DataCollectorImplWrapper::InitializeNetworkMonitoring(bool is_failures_count, bool is_established_count, bool is_reset_count)
 {
 	try
 	{
-		impl->InitializeAllDisksMonitoring(is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght);
+		impl->InitializeNetworkMonitoring(is_failures_count, is_established_count, is_reset_count);
+	}
+	catch (System::Exception^ ex)
+	{
+		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
+	}
+}
+
+void hsm_wrapper::DataCollectorImplWrapper::InitializeDiskMonitoring(const std::string& target, bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght, bool is_average_speed)
+{
+	try
+	{
+		impl->InitializeDiskMonitoring(target, is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght, is_average_speed);
+	}
+	catch (System::Exception^ ex)
+	{
+		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
+	}
+}
+
+void hsm_wrapper::DataCollectorImplWrapper::InitializeAllDisksMonitoring(bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght, bool is_average_speed)
+{
+	try
+	{
+		impl->InitializeAllDisksMonitoring(is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght, is_average_speed);
 	}
 	catch (System::Exception^ ex)
 	{
@@ -576,7 +320,6 @@ hsm_wrapper::HSMSensor<std::string> DataCollectorImplWrapper::CreateStringSensor
 	}
 }
 
-#ifdef ENABLE_OBSOLETE
 HSMLastValueSensor<bool> DataCollectorImplWrapper::CreateLastValueBoolSensor(const std::string& path, bool default_value, const std::string& description)
 {
 	try
@@ -624,7 +367,7 @@ HSMLastValueSensor<int> DataCollectorImplWrapper::CreateLastValueIntSensor(const
 		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
 	}
 }
-#endif
+
 HSMBarSensor<int> DataCollectorImplWrapper::CreateIntBarSensor(const std::string& path, int timeout, int small_period, const std::string& description)
 {
 	try
@@ -674,7 +417,7 @@ HSMBarSensor<double> DataCollectorImplWrapper::CreateDoubleBarSensor(const std::
 		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
 	}
 }
-#ifdef ENABLE_OBSOLETE
+
 template<class T>
 std::shared_ptr<HSMNoParamsFuncSensorImplWrapper<T>> DataCollectorImplWrapper::CreateNoParamsFuncSensor(const std::string& path, const std::string& description, 
 	std::function<T()> function, const std::chrono::milliseconds& interval)
@@ -702,7 +445,7 @@ shared_ptr<HSMParamsFuncSensorImplWrapper<T, U>> DataCollectorImplWrapper::Creat
 		throw std::exception(msclr::interop::marshal_as<std::string>(ex->Message).c_str());
 	}
 }
-#endif
+
 
 hsm_wrapper::DataCollectorProxy::DataCollectorProxy(const std::string& product_key, const std::string& address, int port, const std::string& module) 
 	: impl_wrapper(new DataCollectorImplWrapper(product_key, address, port, module))
@@ -729,29 +472,34 @@ void DataCollectorProxy::StartAsync()
 	impl_wrapper->StartAsync();
 }
 
-void hsm_wrapper::DataCollectorProxy::InitializeSystemMonitoring(bool is_cpu, bool is_free_ram)
+void DataCollectorProxy::StopAsync()
 {
-	impl_wrapper->InitializeSystemMonitoring(is_cpu, is_free_ram);
+	impl_wrapper->StopAsync();
 }
 
-void DataCollectorProxy::InitializeDiskMonitoring(const string& target, bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
+void hsm_wrapper::DataCollectorProxy::InitializeSystemMonitoring(bool is_cpu /*= true*/, bool is_free_ram /*= true*/, bool is_time_in_gc /*= true*/)
 {
-	impl_wrapper->InitializeDiskMonitoring(target, is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght);
+	impl_wrapper->InitializeSystemMonitoring(is_cpu, is_free_ram, is_time_in_gc);
 }
 
-void hsm_wrapper::DataCollectorProxy::InitializeAllDisksMonitoring(bool is_free_space, bool is_free_space_prediction, bool is_active_time, bool is_queue_lenght)
+void hsm_wrapper::DataCollectorProxy::InitializeDiskMonitoring(const std::string& target, bool is_free_space /*= true*/, bool is_free_space_prediction /*= true*/, bool is_active_time /*= true*/, bool is_queue_lenght /*= true*/, bool is_average_speed /*= true*/)
 {
-	impl_wrapper->InitializeAllDisksMonitoring(is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght);
+	impl_wrapper->InitializeDiskMonitoring(target, is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght, is_average_speed);
 }
 
-void hsm_wrapper::DataCollectorProxy::InitializeProcessMonitoring(bool is_cpu, bool is_memory, bool is_threads)
+void hsm_wrapper::DataCollectorProxy::InitializeAllDisksMonitoring(bool is_free_space /*= true*/, bool is_free_space_prediction /*= true*/, bool is_active_time /*= true*/, bool is_queue_lenght /*= true*/, bool is_average_speed /*= true*/)
 {
-	impl_wrapper->InitializeProcessMonitoring(is_cpu, is_memory, is_threads);
+	impl_wrapper->InitializeAllDisksMonitoring(is_free_space, is_free_space_prediction, is_active_time, is_queue_lenght, is_average_speed);
 }
 
-void hsm_wrapper::DataCollectorProxy::InitializeOsMonitoring(bool is_last_update, bool is_last_restart)
+void hsm_wrapper::DataCollectorProxy::InitializeProcessMonitoring(bool is_cpu /*= true*/, bool is_memory /*= true*/, bool is_threads /*= true*/, bool is_time_in_gc /*= true*/)
 {
-	impl_wrapper->InitializeOsMonitoring(is_last_update, is_last_restart);
+	impl_wrapper->InitializeProcessMonitoring(is_cpu, is_memory, is_threads, is_time_in_gc);
+}
+
+void hsm_wrapper::DataCollectorProxy::InitializeOsMonitoring(bool is_last_update /*= true*/, bool is_last_restart /*= true*/, bool is_version /*= true*/)
+{
+	impl_wrapper->InitializeOsMonitoring(is_last_update, is_last_restart, is_version);
 }
 
 void DataCollectorProxy::InitializeOsLogsMonitoring(bool is_warnig, bool is_error)
@@ -774,9 +522,19 @@ void hsm_wrapper::DataCollectorProxy::InitializeProductVersion(const std::string
 	impl_wrapper->InitializeProductVersion(version);
 }
 
-void hsm_wrapper::DataCollectorProxy::InitializeCollectorMonitoring(bool is_alive, bool version, bool status)
+void DataCollectorProxy::InitializeQueueDiagnostic(bool is_overflow /*= true*/, bool is_process_time /*= true*/, bool is_values_count /*= true*/, bool is_content_size /*= true*/)
 {
-	impl_wrapper->InitializeCollectorMonitoring(is_alive, version, status);
+	impl_wrapper->InitializeQueueDiagnostic(is_overflow, is_process_time, is_values_count, is_content_size);
+}
+
+void hsm_wrapper::DataCollectorProxy::InitializeCollectorMonitoring(bool is_alive /*= true*/, bool version /*= true*/, bool is_errors /*= true*/)
+{
+	impl_wrapper->InitializeCollectorMonitoring(is_alive, version, is_errors);
+}
+
+void DataCollectorProxy::InitializeNetworkMonitoring(bool is_failures_count /*= true*/, bool is_established_count /*= true*/, bool is_reset_count /*= true*/)
+{
+	impl_wrapper->InitializeNetworkMonitoring(is_failures_count, is_established_count, is_reset_count);
 }
 
 BoolSensor DataCollectorProxy::CreateBoolSensor(const std::string& path, const std::string& description)
@@ -819,7 +577,6 @@ StringSensor DataCollectorProxy::CreateStringSensor(const std::string& path, con
 	return impl_wrapper->CreateStringSensor(path, options);
 }
 
-#ifdef ENABLE_OBSOLETE
 BoolLastValueSensor DataCollectorProxy::CreateLastValueBoolSensor(const std::string& path, bool default_value, const std::string& description)
 {
 	return impl_wrapper->CreateLastValueBoolSensor(path, default_value, description);
@@ -839,7 +596,7 @@ StringLastValueSensor DataCollectorProxy::CreateLastValueStringSensor(const std:
 {
 	return impl_wrapper->CreateLastValueStringSensor(path, default_value, description);
 }
-#endif
+
 
 IntBarSensor DataCollectorProxy::CreateIntBarSensor(const std::string& path, int timeout, int small_period, const std::string& description)
 {
@@ -861,8 +618,6 @@ DoubleBarSensor DataCollectorProxy::CreateDoubleBarSensor(const std::string& pat
 {
 	return impl_wrapper->CreateDoubleBarSensor(path, options);
 }
-
-#ifdef ENABLE_OBSOLETE
 
 #define InstantiateOneParamTemplates(X)\
 template shared_ptr<HSMNoParamsFuncSensorImplWrapper<X>> DataCollectorImpl::CreateNoParamsFuncSensor<X>\
@@ -901,4 +656,3 @@ InstantiateTwoParamTemplates(string, double)
 InstantiateTwoParamTemplates(string, bool)
 InstantiateTwoParamTemplates(string, string)
 
-#endif
